Fix pre-existing build errors in batch.rs and lockfree_queue.rs
Remove duplicate method definitions (len, is_empty, ops) in shell batch
module, and fix Vec<i32> vs Vec<usize> type mismatch in lockfree queue
test.

+++ crates/liquide-session/src/desktop/lockfree_queue.rs
@@ -532,7 +532,7 @@ mod tests {
         // might interleave — within each batch, FIFO is guaranteed).
         let mut sorted = collected;
         sorted.sort();
-        let expected: Vec<i32> = (0..total_items as i32).collect();
+        let expected: Vec<usize> = (0..total_items).collect();
         assert_eq!(sorted, expected);
     }
